+++ lnst/Recipes/ENRT/ConfigMixins/CoalescingHWConfigMixin.py
@@ -1,61 +1,101 @@
-from lnst.Common.Parameters import BoolParam
+from copy import copy
 
+from lnst.Common.Parameters import DictParam
 from lnst.Recipes.ENRT.ConfigMixins.BaseHWConfigMixin import BaseHWConfigMixin
 
 
 class CoalescingHWConfigMixin(BaseHWConfigMixin):
     """
     This class is an extension to the :any:`BaseEnrtRecipe` class that enables
-    adaptive coalescing configuration on the devices defined by
+    interrupt coalescing configuration on the devices defined by
     :attr:`coalescing_hw_config_dev_list` property.
 
-    :param adaptive_tx_coalescing:
-        (optional test parameter) boolean to enable/disable TX adaptive
-        coalescing on the devices
-    :param adaptive_rx_coalescing:
-        (optional test parameter) boolean to enable/disable RX adaptive
-        coalescing on the devices
+    This mixin is a "multi device" variant of the :any:`CoalescingHWConfigMixin`
+    and allows configuration of individual devices instead of sharing the same
+    configuration. This may be required when different interrupt coalescing is
+    required by a flow generator and receiver host.
+
+    :param coalescing_settings:
+        (optional test parameter) dictionary to specify coalescing settings
+        for individual devices, in 'ethtool -C ...' format
     """
-
-    adaptive_rx_coalescing = BoolParam(mandatory=False)
-    adaptive_tx_coalescing = BoolParam(mandatory=False)
-
-    @property
-    def coalescing_hw_config_dev_list(self):
-        """
-        The value of this property is a list of devices for which the
-        adaptive coalescing features should be configured. It has to be
-        defined by a derived class.
-        """
-        return []
+    coalescing_settings = DictParam(mandatory=False, default={})
 
     def hw_config(self, config):
         super().hw_config(config)
 
-        for param in ["adaptive_rx_coalescing", "adaptive_tx_coalescing"]:
-            param_value = getattr(self.params, param, None)
-            if param_value is not None:
+        device_settings = self._parse_device_settings(self.params.coalescing_settings)
+        for device, device_setting in device_settings.items():
+            device_setting_copy = copy(device_setting)
+            # first, fetch adaptive setting as it needs to be turned off before
+            # configuring individual coalescing settings
+            for param in ["adaptive-tx", "adaptive-rx"]:
+                if param_value := device_setting_copy.pop(param, None):
+                    self._configure_dev_attribute(
+                        config,
+                        [device],
+                        coalescing_param_to_device_attribute(param),
+                        param_value == "on",
+                    )
+
+            for param, param_value in device_setting_copy.items():
                 self._configure_dev_attribute(
                     config,
-                    self.coalescing_hw_config_dev_list,
-                    param,
-                    param_value
+                    [device],
+                    coalescing_param_to_device_attribute(param),
+                    param_value,
                 )
 
     def hw_deconfig(self, config):
-        for feature in ['adaptive_rx_coalescing', 'adaptive_tx_coalescing']:
-            self._deconfigure_dev_attribute(
-                config,
-                self.coalescing_hw_config_dev_list,
-                feature,
-            )
+        device_settings = self._parse_device_settings(self.params.coalescing_settings)
+        for device, device_setting in device_settings.items():
+            device_setting_copy = copy(device_setting)
+            for param in device_setting:
+                if param in ["adaptive-tx", "adaptive-rx"]:
+                    continue
+                device_setting_copy.pop(param)
+                self._deconfigure_dev_attribute(
+                    config,
+                    [device],
+                    coalescing_param_to_device_attribute(param),
+                )
+
+            for param in device_setting_copy:
+                self._deconfigure_dev_attribute(
+                    config,
+                    [device],
+                    coalescing_param_to_device_attribute(param),
+                )
 
         super().hw_deconfig(config)
 
     def describe_hw_config(self, config):
         desc = super().describe_hw_config(config)
-        for param in ["adaptive_rx_coalescing", "adaptive_tx_coalescing"]:
+
+        device_settings = self._parse_device_settings(self.params.coalescing_settings)
+        coalescing_attrs = {
+            coalescing_param_to_device_attribute(key)
+            for setting in device_settings.values()
+            for key in setting.keys()
+        }
+        for attr in sorted(coalescing_attrs):
             desc.extend(
-                self._describe_dev_attribute(config, param)
+                self._describe_dev_attribute(
+                    config,
+                    attr,
+                )
             )
+
         return desc
+
+
+def coalescing_param_to_device_attribute(param):
+    """
+    This mixin accepts the coalesce_settings keys in 'ethtool -C' format
+    The Device class uses more descriptive property names along with
+    underscores, so we need to translate the keys to property names
+    """
+    if param in ["adaptive-tx", "adaptive-rx"]:
+        return param.replace("-", "_") + "_coalescing"
+
+    return "coalescing_" + param.replace("-", "_")
